+++ src/siun/state.py
@@ -4,7 +4,6 @@ import datetime
 import importlib.util
 import shutil
 import tempfile
-from enum import Enum
 from importlib.machinery import SourceFileLoader
 from pathlib import Path
 from typing import Any
@@ -13,7 +12,7 @@ from pydantic import BaseModel, Field, ValidationError
 
 from siun.criteria import CriterionAvailable, CriterionCount, CriterionCritical, SiunCriterion
 from siun.errors import CriterionError
-from siun.models import V2Threshold
+from siun.models import ClickColor, V2Threshold
 from siun.util import get_default_criteria_dir
 
 BUILTIN_CRITERIA = {
@@ -58,62 +57,6 @@ def _load_user_criteria(*, criteria_settings: dict[str, Any], include_path: Path
     return user_criteria
 
 
-class SortableEnum(Enum):
-    """Sortable Enum subclass."""
-
-    def __le__(self, other: Any):
-        """Make enum sortable: Less Than or Equal."""
-        if self.__class__ is other.__class__:
-            if self == other:
-                return True
-            return self.__lt__(other)
-        return NotImplemented
-
-    def __lt__(self, other: Any):
-        """Make enum sortable: Less Than."""
-        if self.__class__ is other.__class__:
-            return list(self.__class__).index(self) < list(self.__class__).index(other)
-        return NotImplemented
-
-
-class Threshold(SortableEnum):
-    """Threshold levels."""
-
-    available = "available"
-    warning = "warning"
-    critical = "critical"
-
-
-class State(SortableEnum):
-    """Define update state."""
-
-    UNKNOWN = "UNKNOWN"
-    OK = "OK"
-    AVAILABLE_UPDATES = "AVAILABLE_UPDATES"
-    WARNING_UPDATES = "WARNING_UPDATES"
-    CRITICAL_UPDATES = "CRITICAL_UPDATES"
-
-
-class StateText(Enum):
-    """Translate state to text representation."""
-
-    OK = "Ok"
-    AVAILABLE_UPDATES = "Updates available"
-    WARNING_UPDATES = "Updates recommended"
-    CRITICAL_UPDATES = "Updates required"
-    UNKNOWN = "Unknown"
-
-
-class StateColor(Enum):
-    """Translate state to color."""
-
-    OK = "green"
-    AVAILABLE_UPDATES = "blue"
-    WARNING_UPDATES = "yellow"
-    CRITICAL_UPDATES = "red"
-    UNKNOWN = "magenta"
-
-
 class FormatObject(BaseModel):
     """Objects for output formatting."""
 
@@ -136,9 +79,9 @@ class Updates(BaseModel):
     thresholds: list[V2Threshold] = []
     available_updates: list[str] = []
     matched_criteria: dict[str, dict[str, Any]] = {}
-    state: State = State.UNKNOWN
-    last_state: State = State.UNKNOWN
     last_update: datetime.datetime = datetime.datetime.now(tz=datetime.UTC)
+    match: V2Threshold | None = None
+    last_match: V2Threshold | None = None
 
     def touch(self) -> None:
         """Set last_update to now."""
@@ -155,14 +98,18 @@ class Updates(BaseModel):
         return len(self.available_updates)
 
     @property
-    def color(self) -> StateColor:
-        """Get color based on state."""
-        return getattr(StateColor, self.state.name)
+    def color(self) -> ClickColor:
+        """Get color of matched threshold."""
+        if not self.match:
+            return ClickColor.reset
+        return self.match.color
 
     @property
-    def text_value(self) -> StateText:
-        """Get text value based on state."""
-        return getattr(StateText, self.state.name)
+    def text_value(self) -> str:
+        """Get text value of matched threshold."""
+        if not self.match:
+            return "No matches."
+        return self.match.text
 
     @property
     def format_object(self) -> FormatObject:
@@ -173,10 +120,10 @@ class Updates(BaseModel):
             matched_criteria=", ".join(self.matched_criteria.keys()),
             matched_criteria_short=",".join([match[:2] for match in self.matched_criteria]),
             score=self.score,
-            status_text=self.text_value.value,
+            status_text=self.text_value,
             update_count=self.count,
             state_color=self.color.value,
-            state_name=self.text_value.name,
+            state_name=self.text_value,
         )
 
     def update(self, available_updates: list[str] | None = None) -> None:
@@ -211,9 +158,8 @@ class Updates(BaseModel):
 
         for threshold in self.thresholds:
             if self.score >= threshold.score:
-                self.state = State(f"{threshold.name.upper()}_UPDATES")
+                self.match = threshold
                 break
-            self.state = State("OK")
 
     def persist_state(self, state_file_path: Path) -> None:
         """
@@ -227,7 +173,9 @@ class Updates(BaseModel):
             # Create parent dir for state file path if it doesn't exist
             Path.mkdir(Path(state_file_path.parent))
         with tempfile.NamedTemporaryFile(mode="w+") as update_file:
-            update_file.write(self.model_dump_json(exclude={"thresholds"}))
+            update_file.write(
+                self.model_dump_json(exclude={"thresholds", "last_match", "matched_criteria", "criteria_settings"})
+            )
             update_file.flush()
             shutil.copy(update_file.name, state_file_path)